+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/controller/TagController.java
@@ -1,0 +1,48 @@
+package com.itachallenge.challenge.controller;
+
+import com.itachallenge.challenge.dto.GenericResultDto;
+import com.itachallenge.challenge.dto.TagDto;
+import com.itachallenge.challenge.service.ITagService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import reactor.core.publisher.Mono;
+
+import java.util.UUID;
+
+@RestController
+@Validated
+@RequestMapping(value = "/itachallenge/api/v1/tags")
+public class TagController {
+
+    @Autowired
+    private ITagService tagService;
+
+    @GetMapping("/{languageId}")
+    @Operation(
+            operationId = "Get tags by languageId",
+            summary = "Get all tags filtered by languageId.",
+            description = "Retrieve all tags that match the specified languageId.",
+            responses = {
+                    @ApiResponse(responseCode = "200", content = {@Content(schema = @Schema(implementation = GenericResultDto.class), mediaType = "application/json")}),
+                    @ApiResponse(responseCode = "404", description = "No tags found for the specified languageId."),
+                    @ApiResponse(responseCode = "400", description = "Malformed or invalid parameter(s)."),
+                    @ApiResponse(responseCode = "500", description = "Internal Server Error")
+            }
+    )
+    public Mono<ResponseEntity<GenericResultDto<TagDto>>> getTagsByLanguageId(@PathVariable UUID languageId) {
+        return tagService.getTagsByLanguageId(languageId)
+                .map(ResponseEntity::ok)
+                .switchIfEmpty(Mono.just(ResponseEntity.status(HttpStatus.NOT_FOUND).build()))
+                .onErrorResume(e -> Mono.just(ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build()));
+    }
+}